+++ frontend/src/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from "vue";
 import { createPinia } from "pinia";
+import { router } from "./router";
 import App from "./App.vue";
 import "./assets/css/tailwind.css";
 
@@ -8,4 +9,5 @@ const pinia = createPinia();
 
 const app = createApp(App);
 
-app.use(pinia).mount("#app");
+app.use(pinia).use(router);
+app.mount("#app");
